Reformats and changes names of a few files
