+++ app/src/main/java/com/example/silentguardian_android/fragments/InsertThresholdMessageDialogFragment.java
@@ -1,4 +1,4 @@
-package com.example.silentguardian_android;
+package com.example.silentguardian_android.fragments;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -15,6 +15,7 @@ import androidx.annotation.Nullable;
 import androidx.fragment.app.DialogFragment;
 
 import com.example.silentguardian_android.Database.SharePreferenceHelper;
+import com.example.silentguardian_android.R;
 
 public class InsertThresholdMessageDialogFragment extends DialogFragment {
