--- app/src/main/res/layout/activity_add_contact_to_threshload.xml
@@ -1,66 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".AddContactToThreshload">
-
-    <TextView
-        android:id="@+id/addNameThresholdTV"
-        android:layout_width="304dp"
-        android:layout_height="34dp"
-        android:layout_marginTop="161dp"
-        android:layout_marginBottom="100dp"
-        android:gravity="center_vertical"
-        android:text="Name"
-        android:textAlignment="center"
-        app:layout_constraintBottom_toTopOf="@+id/phoneNumberThresholdEditText"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <TextView
-        android:id="@+id/phoneNumberThresholdEditText"
-        android:layout_width="277dp"
-        android:layout_height="60dp"
-        android:layout_marginBottom="108dp"
-        android:gravity="center_vertical"
-        android:text="@android:string/emptyPhoneNumber"
-        android:textAlignment="center"
-        app:layout_constraintBottom_toTopOf="@+id/addContactToThresholdButton"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/addNameThresholdTV" />
-
-    <Button
-        android:id="@+id/addContactToThresholdButton"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentLeft="true"
-        android:layout_alignParentBottom="true"
-        android:layout_centerHorizontal="true"
-        android:layout_marginBottom="69dp"
-        android:text="Add contact to threshold"
-        app:layout_constraintBottom_toTopOf="@+id/cancelThresholdButton"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/phoneNumberThresholdEditText" />
-
-    <Button
-        android:id="@+id/cancelThresholdButton"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentEnd="true"
-        android:layout_alignParentRight="true"
-        android:layout_alignParentBottom="true"
-        android:layout_centerHorizontal="true"
-        android:layout_marginBottom="107dp"
-        android:text="@android:string/cancel"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/addContactToThresholdButton" />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
